+++ talk/ics/sdk/base/win/h264_video_mft_encoder.cpp
@@ -450,8 +450,8 @@ int H264VideoMFTEncoder::Encode(
       }
       scPosition += naluPosition;
       scPositions[scPositionsLength++] = scPosition;
-      scLengths[scPositionsLength-1] = scLength;
-      scPosition += scLength;
+      scLengths[scPositionsLength-1] = static_cast<int32_t>(scLength);
+      scPosition += static_cast<int32_t>(scLength);
     }
     if (scPositionsLength == 0) {
       LOG(LS_ERROR) << "Start code is not found for H264 codec!";
@@ -624,8 +624,8 @@ int H264VideoMFTEncoder::EncodeOnEncoderThread(const webrtc::VideoFrame& input_i
         }
         scPosition += naluPosition;
         scPositions[scPositionsLength++] = scPosition;
-        scLengths[scPositionsLength - 1] = scLength;
-        scPosition += scLength;
+        scLengths[scPositionsLength - 1] = static_cast<int32_t>(scLength);
+        scPosition += static_cast<int32_t>(scLength);
     }
     if (scPositionsLength == 0) {
         LOG(LS_ERROR) << "Start code is not found for H264 codec!";
